Refactored customUserAgent and instanceLock
